requirements.txt
Django==4.2.x
djangorestframework==3.x.x
mysqlclient==2.x.x